devpulse-backend/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.1,<1.0.0
httpx>=0.27.0,<1.0.0
pydantic>=2.7.1,<3.0.0
python-dotenv>=1.0.1,<2.0.0
slowapi>=0.1.9,<1.0.0
pyjwt>=2.8.0,<3.0.0
bcrypt>=4.1.0,<5.0.0
websockets>=12.0,<14.0
pyyaml>=6.0,<7.0

# PostgreSQL + SQLAlchemy (primary data layer)
asyncpg>=0.29.0,<1.0.0
sqlalchemy[asyncio]>=2.0.30,<3.0.0
alembic>=1.13.0,<2.0.0

# SQLite (test-only fallback)
aiosqlite>=0.19.0,<1.0.0

# Redis caching (with C parser)
redis>=5.0.0,<8.0.0
hiredis>=3.0.0,<4.0.0

# Billing
stripe>=8.0.0,<10.0.0

# Email
sendgrid>=6.11.0,<7.0.0

# Analytics & Monitoring
posthog>=3.3.0,<4.0.0
sentry-sdk[fastapi]>=1.40.0,<3.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
pytest-cov>=5.0.0,<6.0.0
